requirements.txt
dipy
numpy
matplotlib
pandas
fury
scipy.ndimage